Refactor "Recursion"
