+++ modules.json
@@ -6,6 +6,6 @@
     { "name": "case_cond_and_if", "modules": ["api", "nums"] },
     { "name": "anonymous_functions", "modules": ["logger"] },
     { "name": "binaries_strings_and_charlists", "modules": ["str", "secret"] },
-    { "name": "recursion", "modules": ["recursion"] }
+    { "name": "recursion", "modules": ["stats", "interview"] }
   ]
 }
